feat(EscCredentialProvider): Adding missing tests to verify our changes.

+++ src/Credentials/EcsCredentialProvider.php
@@ -55,7 +55,7 @@ class EcsCredentialProvider
             : (getenv(self::ENV_TIMEOUT) ?: self::DEFAULT_ENV_TIMEOUT);
         $this->retries = (int) isset($config['retries'])
             ? $config['retries']
-            : (getenv(self::ENV_RETRIES) ?: self::DEFAULT_ENV_RETRIES);
+            : ((int) getenv(self::ENV_RETRIES) ?: self::DEFAULT_ENV_RETRIES);
         $this->attempts = 0;
 
         $this->client = $config['client'] ?? \Aws\default_http_handler();
